Extract formatPromotedCommits to its own file
Just a move in this commit. No code changes.
